Fixing limits V2

+++ src/test/java/fr/dawan/formation/testunitaire/Calculatrice.java
@@ -2,9 +2,9 @@ package fr.dawan.formation.testunitaire;
 
 public class Calculatrice {
 
-	public static int additionne(int value1, int value2) {
+	public static long additionne(int value1, int value2) {
 		
-		return value1+value2;
+		return (long)value1 +value2; // caster l'une des deux valeurs, pas les 2.
 	}
 
 }
